Invalidate metadata cache when schema change happens in SQL API
