+++ core/src/main/java/com/scalar/db/sql/DdlStatementExecutor.java
@@ -4,6 +4,9 @@ import com.scalar.db.api.DistributedTransactionAdmin;
 import com.scalar.db.api.Scan;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.sql.exception.SqlException;
+import com.scalar.db.sql.metadata.CachedMetadata;
+import com.scalar.db.sql.metadata.CachedNamespaceMetadata;
+import com.scalar.db.sql.metadata.Metadata;
 import com.scalar.db.sql.statement.CreateCoordinatorTableStatement;
 import com.scalar.db.sql.statement.CreateIndexStatement;
 import com.scalar.db.sql.statement.CreateNamespaceStatement;
@@ -23,9 +26,11 @@ import javax.annotation.concurrent.ThreadSafe;
 public class DdlStatementExecutor implements DdlStatementVisitor<Void, Void> {
 
   private final DistributedTransactionAdmin admin;
+  private final Metadata metadata;
 
-  DdlStatementExecutor(DistributedTransactionAdmin admin) {
+  DdlStatementExecutor(DistributedTransactionAdmin admin, Metadata metadata) {
     this.admin = Objects.requireNonNull(admin);
+    this.metadata = Objects.requireNonNull(metadata);
   }
 
   public void execute(DdlStatement statement) {
@@ -53,6 +58,13 @@ public class DdlStatementExecutor implements DdlStatementVisitor<Void, Void> {
           tableMetadata,
           statement.ifNotExists,
           statement.options);
+
+      // Invalidate the metadata cache
+      metadata
+          .getNamespace(statement.namespaceName)
+          .filter(n -> n instanceof CachedNamespaceMetadata)
+          .map(n -> (CachedNamespaceMetadata) n)
+          .ifPresent(CachedNamespaceMetadata::invalidateTableNamesCache);
       return null;
     } catch (ExecutionException e) {
       throw new SqlException("Failed to create a table", e);
@@ -115,6 +127,11 @@ public class DdlStatementExecutor implements DdlStatementVisitor<Void, Void> {
         }
       }
       admin.dropNamespace(statement.namespaceName, statement.ifExists);
+
+      // Invalidate the metadata cache
+      if (metadata instanceof CachedMetadata) {
+        ((CachedMetadata) metadata).invalidateCache(statement.namespaceName);
+      }
       return null;
     } catch (ExecutionException e) {
       throw new SqlException("Failed to drop a namespace", e);
@@ -125,6 +142,17 @@ public class DdlStatementExecutor implements DdlStatementVisitor<Void, Void> {
   public Void visit(DropTableStatement statement, Void context) {
     try {
       admin.dropTable(statement.namespaceName, statement.tableName, statement.ifExists);
+
+      // Invalidate the metadata cache
+      metadata
+          .getNamespace(statement.namespaceName)
+          .filter(n -> n instanceof CachedNamespaceMetadata)
+          .map(n -> (CachedNamespaceMetadata) n)
+          .ifPresent(
+              n -> {
+                n.invalidateTableNamesCache();
+                n.invalidateTableMetadataCache(statement.tableName);
+              });
       return null;
     } catch (ExecutionException e) {
       throw new SqlException("Failed to drop a table", e);
@@ -180,6 +208,13 @@ public class DdlStatementExecutor implements DdlStatementVisitor<Void, Void> {
           statement.columnName,
           statement.ifNotExists,
           statement.options);
+
+      // Invalidate the metadata cache
+      metadata
+          .getNamespace(statement.namespaceName)
+          .filter(n -> n instanceof CachedNamespaceMetadata)
+          .map(n -> (CachedNamespaceMetadata) n)
+          .ifPresent(n -> n.invalidateTableMetadataCache(statement.tableName));
       return null;
     } catch (ExecutionException e) {
       throw new SqlException("Failed to create a index table", e);
@@ -191,6 +226,13 @@ public class DdlStatementExecutor implements DdlStatementVisitor<Void, Void> {
     try {
       admin.dropIndex(
           statement.namespaceName, statement.tableName, statement.columnName, statement.ifExists);
+
+      // Invalidate the metadata cache
+      metadata
+          .getNamespace(statement.namespaceName)
+          .filter(n -> n instanceof CachedNamespaceMetadata)
+          .map(n -> (CachedNamespaceMetadata) n)
+          .ifPresent(n -> n.invalidateTableMetadataCache(statement.tableName));
       return null;
     } catch (ExecutionException e) {
       throw new SqlException("Failed to drop a index table", e);
