+++ core/src/main/java/com/scalar/db/sql/TransactionSession.java
@@ -36,7 +36,7 @@ public class TransactionSession implements SqlStatementSession {
     this.metadata = Objects.requireNonNull(metadata);
     statementValidator = new StatementValidator(metadata);
     dmlStatementExecutor = new DmlStatementExecutor(metadata);
-    ddlStatementExecutor = new DdlStatementExecutor(admin);
+    ddlStatementExecutor = new DdlStatementExecutor(admin, metadata);
   }
 
   @VisibleForTesting
