+++ core/src/main/java/com/scalar/db/sql/metadata/NamespaceMetadata.java
@@ -1,129 +1,13 @@
 package com.scalar.db.sql.metadata;
 
-import com.google.common.base.MoreObjects;
-import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
-import com.google.common.collect.ImmutableList;
-import com.scalar.db.api.DistributedTransactionAdmin;
-import com.scalar.db.exception.storage.ExecutionException;
-import com.scalar.db.sql.exception.SqlException;
-import java.util.ArrayList;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-import javax.annotation.Nonnull;
-import javax.annotation.concurrent.ThreadSafe;
 
-@ThreadSafe
-public class NamespaceMetadata {
+public interface NamespaceMetadata {
 
-  private final String name;
+  String getName();
 
-  private final LoadingCache<String, ImmutableList<String>> tableNamesCache;
-  private final LoadingCache<String, Optional<TableMetadata>> tableMetadataCache;
+  Map<String, TableMetadata> getTables();
 
-  NamespaceMetadata(String name, DistributedTransactionAdmin admin, long cacheExpirationTimeSecs) {
-    this.name = Objects.requireNonNull(name);
-
-    tableNamesCache = createTableNamesCache(admin, cacheExpirationTimeSecs);
-    tableMetadataCache = createTableMetadataCache(admin, cacheExpirationTimeSecs);
-  }
-
-  private LoadingCache<String, ImmutableList<String>> createTableNamesCache(
-      DistributedTransactionAdmin admin, long cacheExpirationTimeSecs) {
-    CacheBuilder<Object, Object> builder = CacheBuilder.newBuilder();
-    if (cacheExpirationTimeSecs > 0) {
-      builder.expireAfterWrite(cacheExpirationTimeSecs, TimeUnit.SECONDS);
-    }
-    return builder.build(
-        new CacheLoader<String, ImmutableList<String>>() {
-          @Override
-          public ImmutableList<String> load(@Nonnull String namespaceName) {
-            try {
-              return ImmutableList.copyOf(new ArrayList<>(admin.getNamespaceTableNames(name)));
-            } catch (ExecutionException e) {
-              throw new SqlException("Failed to check the namespace existence", e);
-            }
-          }
-        });
-  }
-
-  private LoadingCache<String, Optional<TableMetadata>> createTableMetadataCache(
-      DistributedTransactionAdmin admin, long cacheExpirationTimeSecs) {
-    CacheBuilder<Object, Object> builder = CacheBuilder.newBuilder();
-    if (cacheExpirationTimeSecs > 0) {
-      builder.expireAfterWrite(cacheExpirationTimeSecs, TimeUnit.SECONDS);
-    }
-    return builder.build(
-        new CacheLoader<String, Optional<TableMetadata>>() {
-          @Override
-          public Optional<TableMetadata> load(@Nonnull String tableName) {
-            try {
-              com.scalar.db.api.TableMetadata tableMetadata =
-                  admin.getTableMetadata(name, tableName);
-              if (tableMetadata == null) {
-                return Optional.empty();
-              }
-              return Optional.of(new TableMetadata(name, tableName, tableMetadata));
-            } catch (ExecutionException e) {
-              throw new SqlException("Failed to get a table metadata", e);
-            }
-          }
-        });
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public Map<String, TableMetadata> getTables() {
-    try {
-      return tableNamesCache.get(name).stream()
-          .map(
-              t -> {
-                try {
-                  return tableMetadataCache.get(t);
-                } catch (java.util.concurrent.ExecutionException e) {
-                  throw (RuntimeException) e.getCause();
-                }
-              })
-          .filter(Optional::isPresent)
-          .collect(Collectors.toMap(t -> t.get().getName(), Optional::get));
-    } catch (java.util.concurrent.ExecutionException e) {
-      throw (RuntimeException) e.getCause();
-    }
-  }
-
-  public Optional<TableMetadata> getTable(String tableName) {
-    try {
-      return tableMetadataCache.get(tableName);
-    } catch (java.util.concurrent.ExecutionException e) {
-      throw (RuntimeException) e.getCause();
-    }
-  }
-
-  @Override
-  public String toString() {
-    return MoreObjects.toStringHelper(this).add("name", name).toString();
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (!(o instanceof NamespaceMetadata)) {
-      return false;
-    }
-    NamespaceMetadata that = (NamespaceMetadata) o;
-    return Objects.equals(name, that.name);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(name);
-  }
+  Optional<TableMetadata> getTable(String tableName);
 }
